+++ editor/package.json
@@ -53,6 +53,7 @@
     "react-json-tree": "^0.15.0",
     "react-resizable": "^3.0.1",
     "react-spring": "^9.3.2",
+    "react-textarea-autosize": "^8.3.4",
     "react-use-measure": "^2.1.1",
     "recoil": "^0.2.0",
     "rxdb": "^10.5.4",
